isl_multi_templ.c: extract out isl_multi_tuple_id_templ.c
Although these functions could make sense for isl_multi_id,
there is no need to add them at first.
