+++ include/isl/val.h
@@ -17,6 +17,7 @@ ISL_DECLARE_MULTI_ARITH(val)
 ISL_DECLARE_MULTI_NAN(val)
 ISL_DECLARE_MULTI_DIMS(val)
 ISL_DECLARE_MULTI_DIM_ID(val)
+ISL_DECLARE_MULTI_TUPLE_ID(val)
 ISL_DECLARE_MULTI_WITH_DOMAIN(val)
 
 __isl_export
