+++ isl_aff.c
@@ -3906,6 +3906,7 @@ error:
 #include <isl_multi_nan_templ.c>
 #include <isl_multi_product_templ.c>
 #include <isl_multi_splice_templ.c>
+#include <isl_multi_tuple_id_templ.c>
 #include <isl_multi_zero_templ.c>
 
 /* Construct an isl_multi_aff living in "space" that corresponds
@@ -6432,6 +6433,7 @@ error:
 #include <isl_multi_nan_templ.c>
 #include <isl_multi_product_templ.c>
 #include <isl_multi_splice_templ.c>
+#include <isl_multi_tuple_id_templ.c>
 #include <isl_multi_zero_templ.c>
 
 /* Does "mpa" have a non-trivial explicit domain?
@@ -8304,6 +8306,7 @@ error:
 #include <isl_multi_intersect.c>
 #include <isl_multi_nan_templ.c>
 #include <isl_multi_param_templ.c>
+#include <isl_multi_tuple_id_templ.c>
 
 /* Does "mupa" have a non-trivial explicit domain?
  *
