+++ isl_multi_templ.c
@@ -146,30 +146,6 @@ unsigned FN(MULTI(BASE),dim)(__isl_keep MULTI(BASE) *multi,
 	return multi ? isl_space_dim(multi->space, type) : 0;
 }
 
-const char *FN(MULTI(BASE),get_tuple_name)(__isl_keep MULTI(BASE) *multi,
-	enum isl_dim_type type)
-{
-	return multi ? isl_space_get_tuple_name(multi->space, type) : NULL;
-}
-
-/* Does the specified tuple have an id?
- */
-isl_bool FN(MULTI(BASE),has_tuple_id)(__isl_keep MULTI(BASE) *multi,
-	enum isl_dim_type type)
-{
-	if (!multi)
-		return isl_bool_error;
-	return isl_space_has_tuple_id(multi->space, type);
-}
-
-/* Return the id of the specified tuple.
- */
-__isl_give isl_id *FN(MULTI(BASE),get_tuple_id)(__isl_keep MULTI(BASE) *multi,
-	enum isl_dim_type type)
-{
-	return multi ? isl_space_get_tuple_id(multi->space, type) : NULL;
-}
-
 /* Return the number of base expressions in "multi".
  */
 int FN(MULTI(BASE),size)(__isl_keep MULTI(BASE) *multi)
@@ -330,63 +306,6 @@ __isl_give MULTI(BASE) *FN(MULTI(BASE),reset_space)(
 
 	domain = isl_space_domain(isl_space_copy(space));
 	return FN(MULTI(BASE),reset_space_and_domain)(multi, space, domain);
-}
-
-__isl_give MULTI(BASE) *FN(MULTI(BASE),set_tuple_name)(
-	__isl_keep MULTI(BASE) *multi, enum isl_dim_type type,
-	const char *s)
-{
-	isl_space *space;
-
-	multi = FN(MULTI(BASE),cow)(multi);
-	if (!multi)
-		return NULL;
-
-	space = FN(MULTI(BASE),get_space)(multi);
-	space = isl_space_set_tuple_name(space, type, s);
-
-	return FN(MULTI(BASE),reset_space)(multi, space);
-}
-
-__isl_give MULTI(BASE) *FN(MULTI(BASE),set_tuple_id)(
-	__isl_take MULTI(BASE) *multi, enum isl_dim_type type,
-	__isl_take isl_id *id)
-{
-	isl_space *space;
-
-	multi = FN(MULTI(BASE),cow)(multi);
-	if (!multi)
-		goto error;
-
-	space = FN(MULTI(BASE),get_space)(multi);
-	space = isl_space_set_tuple_id(space, type, id);
-
-	return FN(MULTI(BASE),reset_space)(multi, space);
-error:
-	isl_id_free(id);
-	return NULL;
-}
-
-/* Drop the id on the specified tuple.
- */
-__isl_give MULTI(BASE) *FN(MULTI(BASE),reset_tuple_id)(
-	__isl_take MULTI(BASE) *multi, enum isl_dim_type type)
-{
-	isl_space *space;
-
-	if (!multi)
-		return NULL;
-	if (!FN(MULTI(BASE),has_tuple_id)(multi, type))
-		return multi;
-
-	multi = FN(MULTI(BASE),cow)(multi);
-	if (!multi)
-		return NULL;
-
-	space = FN(MULTI(BASE),get_space)(multi);
-	space = isl_space_reset_tuple_id(space, type);
-
-	return FN(MULTI(BASE),reset_space)(multi, space);
 }
 
 /* Reset the user pointer on all identifiers of parameters and tuples
